update message list upon submit of message
